requirements.txt
# FastAPI & Server
fastapi==0.123.9
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Image Processing
pillow==12.0.0

# ML Models
torch==2.9.1
torchvision==0.24.1
git+https://github.com/openai/CLIP.git

# Vector Search
faiss-cpu==1.13.1
# faiss-gpu==1.13.1  # Uncomment nếu có GPU

# Database
psycopg2-binary==2.9.10
# psycopg2==2.9.10  # Uncomment nếu build from source

# Utils
numpy==2.3.5
pydantic==2.11.9
python-dotenv==1.0.1